+++ backend/Application/Api/GraphQL/Import/AccountImportHandler.cs
@@ -21,18 +21,21 @@ public class AccountImportHandler
         _writer = accountWriter;
     }
 
-    public async Task AddNewAccounts(AccountInfo[] createdAccounts, DateTimeOffset blockSlotTime)
+    public async Task AddNewAccounts(
+        AccountInfo[] createdAccounts, 
+        DateTimeOffset blockSlotTime, 
+        int blockHeight)
     {
         if (createdAccounts.Length == 0) return;
         
-        var accounts = _changeCalculator.MapCreatedAccounts(createdAccounts, blockSlotTime).ToArray();
+        var accounts = _changeCalculator.MapCreatedAccounts(createdAccounts, blockSlotTime, blockHeight).ToArray();
         await _writer.InsertAccounts(accounts);
 
         foreach (var account in accounts)
             _accountLookup.AddToCache(account.BaseAddress.AsString, account.Id);
     }
 
-    public void HandleAccountUpdates(BlockDataPayload payload, TransactionPair[] transactions, Block block)
+    public AccountBalanceUpdate[] HandleAccountUpdates(BlockDataPayload payload, TransactionPair[] transactions, Block block)
     {
         using var counter = _metrics.MeasureDuration(nameof(AccountImportHandler), nameof(HandleAccountUpdates));
 
@@ -50,5 +53,7 @@ public class AccountImportHandler
         var releaseScheduleItems = _changeCalculator.GetAccountReleaseScheduleItems(transactions);
         if (releaseScheduleItems.Length > 0)
             _writer.InsertAccountReleaseScheduleItems(releaseScheduleItems);
+
+        return balanceUpdates;
     }
 }